test(rename): split test_routes_api.py
into test_routes_api_health.py and test_routes_api_status.py
